backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-telegram-bot==20.7
sqlalchemy==2.0.28
requests==2.31.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.25.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
Pillow==10.2.0
